requirements.txt
# ============================================================
#  Phase 2 + 3 Dependencies — Stock Sentiment Dashboard
# ============================================================

# API server
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Market data & RSS
yfinance>=0.2.36
feedparser>=6.0.11
requests>=2.31.0

# Gemini AI
google-genai>=1.0.0

# Database
psycopg2-binary>=2.9.9

# Config / secrets
python-dotenv>=1.0.1
toml>=0.10.2

# Utilities
tenacity>=8.2.3        # retry logic for API calls
schedule>=1.2.0        # daily job scheduler
